refactor: extract SMTP connection pooling
